fix: Remove scoring from AI Assessment flow completely
- Move scoring types to mvp-planner.ts (only used by MVP Planner)
- Remove totalScore from QuizState and QuizResponse interfaces
- Update email endpoints to not reference scoring
- Update email templates to remove score displays
- AI Assessment now works without any scoring references
- MVP Planner scoring remains intact in separate implementation
